minimarket_ml_system/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
prophet==1.1.4
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
joblib==1.3.2
plotly==5.17.0
requests==2.31.0